// angular-app_new/src/app/model/fanfic.ts
// import {ChapterComponent} from "../components/chapter/chapter.component";

export class Fanfic {
  id: number;
  title: string;
  userId: number;
  description: string;
  genre: string;
  imageUrl = 'http://res.cloudinary.com/soaringbird/image/upload/v1519734001/FreeVector-Book-Icon.jpg';
  tags: string[] = [];
  // chapters: ChapterComponent[] = [];
}